refact: modularization of validate on users, event
